app/requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
flask-marshmallow==0.15.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
tzdata==2023.3
Werkzeug==2.3.4
pandas==1.4.0
